chore:インフラ系のバージョンをlatestではなく固定

+++ internal/infra/repository/task/delete.go
@@ -5,16 +5,18 @@ import (
 	"fmt"
 
 	dtask "github.com/kou-etal/go_todo_app/internal/domain/task"
+	duser "github.com/kou-etal/go_todo_app/internal/domain/user"
 )
 
-func (r *repository) Delete(ctx context.Context, id dtask.TaskID, version uint64) error {
+func (r *repository) Delete(ctx context.Context, id dtask.TaskID, userID duser.UserID, version uint64) error {
 	const q = `
 DELETE FROM task
 WHERE
   id = ?
+  AND user_id = ?
   AND version = ?;
 `
-	res, err := r.q.ExecContext(ctx, q, id.Value(), version)
+	res, err := r.q.ExecContext(ctx, q, id.Value(), userID.Value(), version)
 
 	if err != nil {
 		return fmt.Errorf("taskrepo delete execute: %w", err)
